backend/crud/requirements.txt
fastapi==0.86.0
pydantic[dotenv]
tortoise-orm==0.19.2
asyncpg==0.27.0
bcrypt==4.0.1
aiofiles==22.1.0
python-multipart==0.0.5
uvicorn==0.19.0
pytest==7.2.0
pytest-asyncio==0.20.1
httpx==0.23.0
Faker==15.3.1
